requirements.txt
jupyter==1.0.0
requests==2.28.2
plotly==5.13.0
pandas==1.5.3
seaborn==0.12.2
Matplotlib==3.6.3
beautifulsoup4==4.11.2
selenium==4.8.0
nltk==3.8.1
googletrans==4.0.0rc1
vaderSentiment==3.3.2
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
wordcloud==1.8.2.2
statsmodels==0.13.5
category-encoders==2.5.1
xgboost==1.7.1
shap==0.41.0
scikit-learn==1.1.3
psycopg2==2.9.5
Flask==2.2.3
Flask-WTF==1.1.1
Flask-Bootstrap4==4.0.2
python-dotenv==1.0.0
pytest==7.2.2